srand, lol

+++ src/game/game.cpp
@@ -6,7 +6,7 @@
 #include "SDL.h"
 #include "game/sprite_font.h"
 #include "game/timer_display.h"
-#include <ctime>
+#include <time.h>
 
 using namespace game;
 using namespace base;
@@ -24,7 +24,7 @@ Game::Game()
     backgroundLayer->sprites.push_back(background_);
     font = ResourceManager::instance().Font("agencyb.ttf");
 
-    srand(time(NULL));
+    srand((u32)time(NULL));
     board_ = new Board(8, 8);
     timer_ = new TimerDisplay(60.f);
     state_ = GameStates::Start;
